inet gws into gateways
